docker/embed-docs/requirements.txt
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.51.0
langchain==0.1.20
langchain-community==0.0.38
langchain-google-vertexai==0.1.3
qdrant-client==1.9.1
pypdf==3.17.4
click==8.1.7
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
kubernetes==28.1.0
